requirements.txt
openai
python-telegram-bot
PyGithub
python-dotenv
urllib3<2.0 